Replace InternalSerializer check with lazy check in ElasticContractResolver
This commit reverts the previous check added in SourceConverter to determine whether it's
the InternalSerializer by fixing a bug in ElasticContractResolver that is intended to perform this
check. The bug is that the check is evaluated too early, before serializers have been assigned
to ConnectionSettings. Now, the check is lazily made.